Production and consumption of Pizza!
Pizza number is checked by the client before picking it. If Pizza is 1 or  greater than 1 then removes the pizza and update the number.
The chef checks the tray for the number of pizza. If it is less than its maximum number it means there is some space in the tray. It updates the tray by implementing pizza by 1.

Ensure appropriate synchronization and mutual exclusion using PThreads mutex locks and conditional variables
Print out the Chef/Client action. For example “Chef produced pizza: 1,  “Tray is full cannot produce”, “Client ate pizza: 1” etc